fix: Move _LogRotateConf parser out of combiner
* Due to an issue found with the httpdconf parser that was in
  the combiner I'm moving the newer logrotate parser out of it's
  combiner as well.
* Added docstring to the moved parser, and added doctest.
